+++ R/corridor.R
@@ -23,7 +23,7 @@
 #'
 #' @return A simple feature geometry representing the river corridor
 #' @export
-get_corridor <- function(
+delineate_corridor <- function(
   network, river_centerline, river_surface, aoi, initial_method = "valley",
   buffer = NULL, dem = NULL, max_iterations = 10, capping_method = "direct"
 ) {