+++ R/delineate.R
@@ -19,7 +19,9 @@
 #'   the available methods
 #' @param angle_threshold Only network edges forming angles above this threshold
 #'   (in degrees) are considered when forming segment edges. See
-#'  [get_segments()]  and [rcoins::stroke()]. Only used if `segments` is TRUE.
+#'  [delineate_segments()] and [rcoins::stroke()]. Only used if `segments` is
+#'  TRUE.
+#' @param corridor Whether to carry out the corridor delineation
 #' @param segments Whether to carry out the corridor segmentation
 #' @param riverspace Whether to carry out the riverspace delineation
 #' @param force_download Download data even if cached data is available
@@ -27,41 +29,49 @@
 #'   dataset (see [get_dem()]). Only relevant if `initial_method` is `"valley"`
 #'   and `dem` is NULL
 #'
-#' @return A simple feature geometry
+#' @return A list with the corridor, segments, and riverspace geometries
 #' @export
-delineate_corridor <- function(
+delineate <- function(
   city_name, river_name, crs = NULL, buffer_distance = 2500,
   initial_method = "valley", initial_buffer = NULL, dem = NULL,
   max_iterations = 10, capping_method = "direct", angle_threshold = 90,
-  segments = FALSE, riverspace = FALSE, force_download = FALSE, ...
+  corridor = TRUE, segments = FALSE, riverspace = FALSE,
+  force_download = FALSE, ...
 ) {
+
+  if (segments && !corridor) stop("Segmentation requires corridor delineation.")
 
   # Retrieve all relevant OSM datasets within the buffer_distance
   osm_data <- get_osmdata(
-    city_name, river_name, buffer_distance, crs, force_download = force_download
+    city_name, river_name, buffer_distance = buffer_distance, crs = crs,
+    force_download = force_download, network = corridor, buildings = riverspace
   )
 
   # Get the bounding box and (if not provided) the CRS
-
   if (is.null(crs)) crs <- get_utm_zone(osm_data$aoi)
 
-  # If using the valley method, and the DEM is not provided, retrieve dataset
-  if (initial_method == "valley" && is.null(dem)) {
-    aoi_buff <- buffer(osm_data$aoi, buffer_distance)
-    dem <- get_dem(aoi_buff, crs = crs, force_download = force_download, ...)
+  if (corridor) {
+
+    # If using the valley method, and the DEM is not provided, retrieve dataset
+    if (initial_method == "valley" && is.null(dem)) {
+      aoi_buff <- buffer(osm_data$aoi, buffer_distance)
+      dem <- get_dem(aoi_buff, crs = crs, force_download = force_download, ...)
+    }
+
+    # Set up the combined street and rail network for the delineation
+    network_edges <- dplyr::bind_rows(osm_data$streets, osm_data$railways)
+    network <- as_network(network_edges)
+
+    # Run the corridor delineation on the spatial network
+    aoi_repr <- reproject(osm_data$aoi, crs)
+    corridor <- delineate_corridor(
+      network, osm_data$river_centerline, osm_data$river_surface, aoi_repr,
+      initial_method = initial_method, buffer = initial_buffer, dem = dem,
+      max_iterations = max_iterations, capping_method = capping_method
+    )
+  } else {
+    corridor <- NULL
   }
-
-  # Set up the combined street and rail network for the delineation
-  network_edges <- dplyr::bind_rows(osm_data$streets, osm_data$railways)
-  network <- as_network(network_edges)
-
-  # Run the corridor delineation on the spatial network
-  aoi_repr <- reproject(osm_data$aoi, crs)
-  corridor <- get_corridor(
-    network, osm_data$river_centerline, osm_data$river_surface, aoi_repr,
-    initial_method = initial_method, buffer = initial_buffer, dem = dem,
-    max_iterations = max_iterations, capping_method = capping_method
-  )
 
   if (segments) {
     # Select the relevant part of the network
@@ -69,17 +79,18 @@ delineate_corridor <- function(
     corridor_buffer <- sf::st_buffer(corridor, buffer_corridor)
     network_filtered <- filter_network(network, corridor_buffer)
 
-    corridor <- get_segments(corridor, network_filtered,
-                             osm_data$river_centerline, angle_threshold)
+    segments <- delineate_segments(corridor, network_filtered,
+                                   osm_data$river_centerline, angle_threshold)
+  } else {
+    segments <- NULL
   }
-  if (riverspace) delineate_riverspace()
-  return(corridor)
-}
 
-#' Delinate the riverspace.
-#'
-#' @return A simple feature geometry
-#' @export
-delineate_riverspace <- function() {
-  stop("Riverspace delineation not yet implemented.")
+  if (riverspace) {
+    riverspace <- delineate_riverspace(osm_data$buildings,
+                                       osm_data$river_surface)
+  } else {
+    riverspace <- NULL
+  }
+
+  list(corridor = corridor, segments = segments, riverspace = riverspace)
 }