+++ R/segments.R
@@ -12,8 +12,8 @@
 #'
 #' @return Segment polygons as a simple feature geometry
 #' @export
-get_segments <- function(corridor, network, river_centerline,
-                         angle_threshold = 90) {
+delineate_segments <- function(corridor, network, river_centerline,
+                               angle_threshold = 90) {
 
   # Find river crossings in the network and build continuous strokes from them
   crossings <- get_intersecting_edges(network, river_centerline, index = TRUE)