+++ data-raw/bucharest.R
@@ -5,13 +5,15 @@ epsg_code <- 32635
 buffer_distance <- 2500
 
 # Fetch the OSM data
-bucharest_osm <- get_osmdata(
-  city_name,
-  river_name,
-  crs = epsg_code,
-  buffer_distance = buffer_distance,
-  force_download = TRUE
-)
+bbox <- get_osm_bb(city_name)
+bbox_expanded <- buffer_bbox(bbox, bbox_buffer)
+bucharest_osm <- get_osmdata(bbox_expanded, city_name, river_name,
+                             crs = epsg_code, force_download = TRUE)
+
+# Add delineation to package data
+bucharest_delineation <- delineate(city_name, river_name, crs = epsg_code,
+                                   corridor = TRUE, segments = TRUE,
+                                   riverspace = TRUE)
 
 # Fix encoding issue in the WKT strings
 fix_wkt_encoding <- function(x) {
@@ -30,3 +32,4 @@ bucharest_dem <- get_dem(aoi_buff, crs = epsg_code, force_download = TRUE) |>
 # Save as package data
 usethis::use_data(bucharest_osm, overwrite = TRUE)
 usethis::use_data(bucharest_dem, overwrite = TRUE)
+usethis::use_data(bucharest_delineation, overwrite = TRUE)